// apps/inatrace-fe/src/api/model/apiAddress.ts
export interface ApiAddress {
  /**
   * address
   */
  address?: string;
  /**
   * city
   */
  city?: string;
  /**
   * state / province / region
   */
  state?: string;
  /**
   * ZIP / postal code / p.p. box
   */
  zip?: string;
  /**
   * Village cell
   */
  cell?: string;
  /**
   * Village sector
   */
  sector?: string;
  /**
   * Village name
   */
  village?: string;
  /**
   * Other/additional address field
   */
  otherAddress?: string;
  /**
   * Honduras farm name
   */
  hondurasFarm?: string;
  /**
   * Honduras village name
   */
  hondurasVillage?: string;
  /**
   * Honduras municipality name
   */
  hondurasMunicipality?: string;
  /**
   * Honduras department name
   */
  hondurasDepartment?: string;
  country?: ApiCountry;
}