// apps/inatrace-fe/src/api/model/apiFacilityTranslation.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * INATrace Services API OpenAPI documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */

/**
 * List of facility translations
 */

export interface ApiFacilityTranslation {
  /**
   * Facility name
   */
  name?: string;
  /**
   * Facility language
   */
  language?: ApiFacilityTranslation.LanguageEnum;
}

/**
 * Namespace for property- and property-value-enumerations of ApiFacilityTranslation.
 */
export namespace ApiFacilityTranslation {
  /**
   * All properties of ApiFacilityTranslation.
   */
  export enum Properties {
    /**
     * Facility name
     */
    name = 'name',
    /**
     * Facility language
     */
    language = 'language',
  }

  /**
   * All possible values of language.
   */
  export enum LanguageEnum {
    EN = 'EN',
    DE = 'DE',
    RW = 'RW',
    ES = 'ES',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiFacilityTranslation',
      vars: [
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'name',
          classname: 'ApiFacilityTranslation',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: true,
          datatypeWithEnum: 'ApiFacilityTranslation.LanguageEnum',
          required: false,
          name: 'language',
          classname: 'ApiFacilityTranslation',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
      ],
      validators: {
        name: [],
        language: [],
      },
    };
  }

  // export const ApiFacilityTranslationValidationScheme = {
  //     validators: [],
  //     fields: {
  //               name: {
  //                   validators: []
  //               },
  //               language: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiFacilityTranslation>;
}